refactor: 패키지와 메서드 이름 변경 및 초기값 설정
- model -> domain
- Controller.calculate() ->  Controller.process()
- Controller.resulting() -> Controller.display()
- customDelimiter = null -> customDelimiter = ""
