+++ src/main/java/calculator/Application.java
@@ -5,9 +5,9 @@ import calculator.controller.Controller;
 public class Application {
     public static void main(String[] args) {
         Controller controller = new Controller();
-        
+
         controller.start();
-        controller.calculate();
-        controller.resulting();
+        controller.process();
+        controller.display();
     }
 }
