+++ src/main/java/calculator/domain/Calculator.java
@@ -1,4 +1,4 @@
-package calculator.model;
+package calculator.domain;
 
 import java.util.Vector;
