+++ src/main/java/calculator/domain/InputValidator.java
@@ -1,4 +1,4 @@
-package calculator.model;
+package calculator.domain;
 
 public class InputValidator {
